# speed up sdk rebuild