--- a/golang/cosmos/package.json
+++ b/golang/cosmos/package.json
@@ -9,6 +9,7 @@
   "engines": {
     "node": "^18.12 || ^20.9"
   },
+  "packageManager": "yarn@1.22.19",
   "scripts": {
     "test": "exit 0",
     "build:all": "make",