--- a/packages/deployment/Makefile
+++ b/packages/deployment/Makefile
@@ -3,9 +3,6 @@ REPOSITORY_SDK = ghcr.io/agoric/agoric-sdk
 SS := ../cosmic-swingset/
 
 TAG := unreleased
-GIT_REVISION := $(shell hash=$$(git rev-parse --short HEAD); \
-	dirty=`git diff --quiet 2>/dev/null || echo -dirty`; \
-	echo "$$hash$$dirty")
 
 default: docker-build
 
@@ -18,7 +15,7 @@ docker-build: docker-build-sdk docker-build-solo \
 
 docker-build-sdk:
 	bargs=`node ../xsnap/src/build.js --show-env | sed -e 's/^/ --build-arg=/'`; \
-	docker build $$bargs --build-arg=GIT_REVISION=$(GIT_REVISION) \
+	docker build $$bargs \
 		-t $(REPOSITORY_SDK):$(TAG) --file=Dockerfile.sdk ../..
 	docker tag $(REPOSITORY_SDK):$(TAG) $(REPOSITORY_SDK):latest