closes: #8848

## Description

Cuts a couple minutes wait off re-builds when JS changed and removes the 1min cost for commits.

Benchmarking with `yarn build:sdk` in a3p-integration after `docker system prune`.

Step | master (sec) | PR (sec)
-- | -- | --
Prune + build | `149.2` | `155.5`
Rebuild | `   .7` | `   .6`
Modify vat + build | `177.0` | ` 73.0`
Rebuild | `   .9` | `   .9`
Commit change + build | ` 70.7` | `   .7`
Rebuild | `   .8` | `   .5`

That 73s after changing a vat is because Yarn classic is being dumb about "building fresh packages" for 48s. That will go away with,
- https://github.com/Agoric/agoric-sdk/issues/451

### Security Considerations

n/a, local

### Scaling Considerations

n/a, local

### Documentation Considerations

should be transparent


### Testing Considerations

Tested locally

### Upgrade Considerations

n/a