@turadg 2024-03-21
mysterious failure with Hardened JS:
```
#46 28.74 START ensure enough IST to install bundles
#46 28.74 { totalSize: 204861, cost: 409.72200000000004 }
#46 28.82 { wantMinted: 410 }
#46 28.91 2024-03-21T20:14:28.509Z block-manager: block 1008 begin
#46 28.91 2024-03-21T20:14:28.513Z block-manager: block 1008 commit
#46 29.92 2024-03-21T20:14:29.516Z block-manager: block 1009 begin
#46 29.92 2024-03-21T20:14:29.520Z block-manager: block 1009 commit
#46 30.60 file:///usr/src/proposals/b:localchain/node_modules/execa/lib/error.js:60
#46 30.60 		error = new Error(message);
#46 30.60 		        ^
#46 30.60 
#46 30.60 Error: Command failed with exit code 255: /usr/src/agoric-sdk/node_modules/.bin/agops vaults open --wantMinted 410 --giveCollateral 69
#46 30.60 (TypeError#1)
#46 30.60 TypeError#1: Cannot assign to read only property 'constructor' of object '[object Object]'
#46 30.60 
#46 30.60     at makeError (file:///usr/src/proposals/b:localchain/node_modules/execa/lib/error.js:60:11)
#46 30.60     at handlePromise (file:///usr/src/proposals/b:localchain/node_modules/execa/index.js:124:26)
#46 30.60     at process.processTicksAndRejections (node:internal/process/task_queues:95:5)
#46 30.60     at async executeCommand (file:///usr/src/proposals/b:localchain/node_modules/@agoric/synthetic-chain/dist/lib/index.js:121:22)
#46 30.60     at async Object.vaults (file:///usr/src/proposals/b:localchain/node_modules/@agoric/synthetic-chain/dist/lib/index.js:191:20)
#46 30.60     at async executeOffer (file:///usr/src/proposals/b:localchain/node_modules/@agoric/synthetic-chain/dist/lib/index.js:369:17)
#46 30.60     at async mintIST (file:///usr/src/proposals/b:localchain/node_modules/@agoric/synthetic-chain/dist/lib/index.js:729:3)
#46 30.60     at async ensureISTForInstall (file:///usr/src/proposals/b:localchain/node_modules/@agoric/synthetic-chain/dist/lib/index.js:824:3)
#46 30.60     at async file:///usr/src/proposals/b:localchain/node_modules/@agoric/synthetic-chain/dist/lib/index.js:1005:5
#46 30.60     at async step (file:///usr/src/proposals/b:localchain/node_modules/@agoric/synthetic-chain/dist/lib/index.js:910:3) {
#46 30.60   shortMessage: 'Command failed with exit code 255: /usr/src/agoric-sdk/node_modules/.bin/agops vaults open --wantMinted 410 --giveCollateral 69',
```