@michaelfig 2024-03-22
> mysterious failure with Hardened JS:

Please set the `LOCKDOWN_OPTIONS` environment variable appropriately to debug these errors caused by the "override mistake":
```
LOCKDOWN_OPTIONS='{"errorTaming":"unsafe","stackFiltering":"verbose","overrideTaming":"severe","overrideDebug":["constructor"]}' yarn --silent agops vaults open --wantMinted 410 --giveCollateral 69
```