@turadg 2024-03-24
new error. maybe due to https://github.com/Agoric/agoric-sdk/pull/9128. WDYT @michaelfig ? 

```
#51 0.355 mv: cannot stat 'node_modules/.cache/agoric/golang-built.sum.new': No such file or directory
#51 0.356 Command: agd version failed with code 1
#51 ERROR: process "/bin/sh -c /usr/src/agoric-sdk/scripts/smoketest-binaries.sh" did not complete successfully: exit code: 1
------
 > [install 16/17] RUN /usr/src/agoric-sdk/scripts/smoketest-binaries.sh:
0.342 Checking agd version
0.349 Creating node_modules/.cache/agoric
0.355 mv: cannot stat 'node_modules/.cache/agoric/golang-built.sum.new': No such file or directory
0.356 Command: agd version failed with code 1
```