@michaelfig 2024-03-24
> new error.
This was an oversight on my part.  The `bin/agd` script should use `mv -f`, not just `mv`, as follows:
```diff
diff --git a/bin/agd b/bin/agd
index e30888281..cc741aa2c 100755
--- a/bin/agd
+++ b/bin/agd
@@ -197,7 +197,7 @@ ${NO_BUILD:-false} || (
     diffsha1 "$sum" "${files[@]}" || { echo "$sum has changed"; $do_not_build; } || {
       rm -f "$sum" "$STAMPS/yarn-built"
       lazy_yarn install
-      mv "$sum.new" "$sum"
+      mv -f "$sum.new" "$sum"
     }
     rm -f "$sum.new"

```
Feel free to include this change as a drive-by.