@turadg 2024-03-24
> Feel free to include this change as a drive-by.

Done but still failing the same way. @michaelfig could I ask you to get `agd` compatible with this Dockerfile again?