@michaelfig 2024-03-24
> Done but still failing the same way. @michaelfig could I ask you to get `agd` compatible with this Dockerfile again?

Yes, I can do that later today.  Do you want me to do it on this PR's branch, or somewhere else?